requirements_CI.txt
black==22.12.0
flake8==4.0.1
flake8-bugbear==22.1.11
flake8-pie==0.15.0
isort==5.9.3
mypy==0.931
pytest~=7.2.1

sqlalchemy-stubs==0.4

